feat:Estruturação de Classes

+++ src/main/java/com/residenciasquad9/demo/domain/repository/ProtocoloRepository.java
@@ -3,6 +3,8 @@ package com.residenciasquad9.demo.domain.repository;
 import com.residenciasquad9.demo.domain.entites.Protocolo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProtocoloRepository extends JpaRepository <Protocolo, Integer> {
-    boolean existsById(int id);
+public interface ProtocoloRepository extends JpaRepository <Protocolo, Long> {
+    boolean existsById(Long id);
+
+
 }
